@Tanner requested changes made
