+++ package.json
@@ -55,6 +55,7 @@
     "open": "^6.3.0",
     "request": "^2.81.0",
     "rimraf": "^2.6.3",
+    "semver": "^6.3.0",
     "xml": "^1.0.1"
   },
   "devDependencies": {
